feat: Add trip creation functionality and invite participants
The code changes include adding functionality to create trips and invite participants. This feature allows users to create new trips and send invitations to other users. The changes also involve exporting the `api` constant from the `axios` module and using it to consume the backend API. Additionally, the code includes the use of a date picker to select the start and end dates of the trip. The component `destination-and-date-header.tsx` has been created, which displays the destination and date information for a trip.